AI-Visual-Defect-Detection-FIXED/requirements.txt
streamlit==1.30.0
torch==2.2.0
torchvision==0.17.0
altair==4.2.2
Pillow
numpy
